boards: qemu_x86_tiny: use custom linker script
This enables qemu_x86_tiny to use its own linker script to
better support demand paging. In the process, a new kconfig
CONFIG_BOARD_QEMU_X86_TINY has to be introduced so that
ZEPHYR_BASE can be used to specify the path to the linker
script. The linker script is based on the generic x86-32
one in include/arch/x86/ia32/linker.ld, with additions to
put symbols into boot and pinned sections.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>