ci: clang: make llvm/clang/lld available in PATH
This is to make ld.lld appears on $PATH, and this is needed
if binutils' ld is to be used for linker. Without ld.lld
in $PATH, clang would complain that it cannot find `ld`.

Note that this is because the package lld-15 does not
install the symlink in /usr/bin/ld.lld. The lld package
provided by Ubuntu 20.04 (used by CI) would pull in
llvm/clang version 10 and would symlink to that instead.
And Ubuntu 22.04 would also need this workaround.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>