toolchain: llvm: use binutils' ld linker by default
LLVM's lld currently has issue linking binaries, especially
with C++. So this makes binutils' ld linker as default, since
it has been proved to be working.

This also changes CONFIG_LLVM_USE_LD to CONFIG_LLVM_USE_LLD.
Since they are of opposite usage, it cannot utilize the usual
deprecation process since that would use "select" to enable
both, which cannot not be done in this case.

Note that CI has been using ld to link binaries, and that's
why we haven't seen any issue there. It is due to lld not
being in PATH, and clang thus defaults to use ld instead.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>