soc: intel_adsp: rework host IPC using IPC message service
This reworks the host IPC driver as a backend of the newly
introduced IPC message service. This is the first step to
rework IPC in SOF (Sound Open Firmware) into using a more
generic IPC API instead of a SoC specific one.

For now, it keeps the old interface to maintain usability
as it is going to be a multiple process to rework IPC
over there.

Also, the structure of the new IPC backend resembles
the SoC specific driver to make it easier to compare
between them at this first iteration. Future optimizations
will probably be needed once we start modifying the SOF
side to utilize the new IPC message service API.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>