soc: intel_adsp/ace: don't use default MMU mappings
This adds the arch default MMU mappings into the SoC mapping
array. This is in preparation of not forcibly doing double
mapping in architecture code as this type of double mapping
should be specified for individual regions. This allows
the SoC layer to fine tune the mappings as not anything
needs to be double-mapped.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>