xtensa: imply CONFIG_ARCH_HAS_RESERVED_PAGE_FRAMES if MMU
The Zephyr linker script usually puts something before
z_mapped_start (where .text is), for example, vecbase vectors.
So we need to reserve those space or else k_mem_map() would be
mapping those which may result in faults.

To avoid that, CONFIG_ARCH_HAS_RESERVED_PAGE_FRAMES needs to be
used. Since it is common when MMU is enabled, we should enable
it by default using imply. All current Xtensa MMU SoCs all have
this selected anyway. Using 'imply' instead of 'select' is to
allow it to be disabled if so desired.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>